--- core/src/main/java/timeboard/core/async/AsyncJobState.java
@@ -1,138 +1,0 @@
-package timeboard.core.async;
-
-/*-
- * #%L
- * core
- * %%
- * Copyright (C) 2019 Timeboard
- * %%
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- * 
- * The above copyright notice and this permission notice shall be included in
- * all copies or substantial portions of the Software.
- * 
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
- * THE SOFTWARE.
- * #L%
- */
-
-import timeboard.core.model.OrganizationEntity;
-
-import javax.persistence.*;
-import java.util.Date;
-
-@Entity
-public class AsyncJobState extends OrganizationEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.AUTO)
-    private long id;
-
-    @Column
-    public Long ownerID;
-
-    @Column
-    @Enumerated(EnumType.STRING)
-    public State state;
-
-    @Column
-    public String title;
-
-    @Column(length=1000)
-    private String error;
-
-    @Column(length=1000)
-    private String result;
-
-    @Temporal(TemporalType.TIMESTAMP)
-    @Column
-    private Date startDate;
-
-    @Temporal(TemporalType.TIMESTAMP)
-    @Column
-    private Date endDate;
-
-    public void setOwnerID(Long ownerID) {
-        this.ownerID = ownerID;
-    }
-
-    public Date getStartDate() {
-        return startDate;
-    }
-
-    public void setStartDate(Date startDate) {
-        this.startDate = startDate;
-    }
-
-    public Date getEndDate() {
-        return endDate;
-    }
-
-    public void setEndDate(Date endDate) {
-        this.endDate = endDate;
-    }
-
-    public long getId() {
-        return id;
-    }
-
-    public void setId(long id) {
-        this.id = id;
-    }
-
-    public long getOwnerID() {
-        return ownerID;
-    }
-
-    public void setOwnerID(long ownerID) {
-        this.ownerID = ownerID;
-    }
-
-    public State getState() {
-        return state;
-    }
-
-    public void setState(State state) {
-        this.state = state;
-    }
-
-    public String getError() {
-        return error;
-    }
-
-    public void setError(String error) {
-        this.error = error;
-    }
-
-    public String getResult() {
-        return result;
-    }
-
-    public void setResult(String result) {
-        this.result = result;
-    }
-
-    public String getTitle() {
-        return title;
-    }
-
-    public void setTitle(String title) {
-        this.title = title;
-    }
-
-    public enum State{
-        IN_PROGRESS,
-        DONE,
-        FAILED
-    }
-}
